requirements.txt
Django==1.6.5
Fabric==1.8.3
M2Crypto==0.22.3
Markdown==2.3.1
MySQL-python==1.2.5
Pillow==2.3.0
South==0.8.4
alioss==0.22.4
amqp==1.4.5
anyjson==0.3.3
billiard==3.3.0.17
celery==3.1.12
cssselect==0.9.1
diff-match-patch==20110725.1
django-adminplus==0.3
django-appconf==0.6
django-celery==3.1.10
django-ckeditor-updated==4.2.8
django-concurrency==0.7.1
django-debug-toolbar==1.2.1
django-extensions==1.3.10
django-filter==0.7
django-imagekit==3.2
django-import-export==0.2.2
django-mathfilters==0.2.1
django-oauth2-provider==0.2.6.1
django-registration==1.0
django-registration-defaults==0.4.4
django-rest-swagger==0.1.14
django-reversion==1.8.1
django-simple-captcha==0.4.2
django-supervisor==0.3.2
django-widget-tweaks==1.3
djangorestframework==2.3.13
drf-wrapper==0.1
dogstatsd-python==0.5.1
ecdsa==0.11
fabric-components==0.1.1
ipython==2.1.0
jsonfield==0.9.20
kombu==3.0.19
lxml==3.3.4
oauthlib==0.6.1
paramiko==1.12.3
pilkit==1.1.6
pinyin==0.2.5
pycrypto==2.6.1
pyjade==2.2.0
pyparsing>=1.5.5
pyquery==1.2.8
python-dateutil==2.2
pytz==2013.9
raven==4.1.1
requests==2.1.0
requests-oauthlib==0.4.0
shortuuid==0.4.0
simplejson==3.3.2
six==1.6.1
sqlparse==0.1.10
tablib==0.10.0
urllib3==1.7.1
wsgiref==0.1.2
httplib2==0.9
suds==0.4
eventlet==0.15.2
gevent==1.0.1
gunicorn==19.1.1